experiments/grokking_hard_indicators_v0_2/code/requirements.txt
numpy>=1.24
pyyaml>=6.0
tqdm>=4.66
# Install PyTorch per your platform. On Apple Silicon:
# https://pytorch.org/get-started/locally/
torch>=2.2

